requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pdfminer3k==1.3.1
pluggy==0.13.1
ply==3.11
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.2
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
